Bluetooth low energy - Data chunking (#945)
All data that are used in BLE data transfer is now split into 200 bytes length data packets with an additional byte at the end that indicates the type of the packet. One indicates that there are more packets in the queue, and zero indicates that this is the last packet. The data is split by the sender, sent one by one, and joined back by the receiver.